simple_project/requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-groq==0.3.2
langchain-openai==0.3.25
python-dotenv==1.1.0
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langchain-chroma==0.2.4
pypdf==5.6.0

